[marathon] Add support for readiness checks.
This change adds support for Marathon readiness checks: If the option is
set on Traefik and applications are found to come with configured
readiness checks, the check results (as conducted by Marathon and
exposed through its API) will be taken into account at deployment times
in order to decide if a task should be taken into load-balancing
rotation or not.

Note that I had to extend the event filter that controls how often the
Marathon provider polls the API so that the readiness checker is able to
pick up all necessary readiness state transitions.